invoice_extractor/requirements.txt
pymysql>=1.0.0
cryptography>=41.0.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=1.26.0
lxml>=5.3.0
mysql-connector-python>=9.0.0
pydantic>=2.3.0
pydantic-settings>=2.2.0
APScheduler>=3.10.0